requirements.txt
Flask
Flask-Cors
Flask-RESTful
Flask-JWT-Extended
Flask-Migrate
SQLAlchemy
Werkzeug
passlib
python-dotenv
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-Migrate
SQLAlchemy
Werkzeug
passlib
python-dotenv
